requirements.txt
groq==0.13.1
langchain_community==0.3.13
langchain_google_genai==2.0.7
langchain_text_splitters==0.3.4
pinecone==5.4.2
pinecone_text==0.9.0
protobuf==5.29.2
PyPDF2==3.0.1
python-dotenv==1.0.1
